stringtheory::interning

Trait Interner

pub trait Interner {
    // Required methods
    fn is_empty(&self) -> bool;
    fn len(&self) -> usize;
    fn len_bytes(&self) -> usize;
    fn capacity_bytes(&self) -> usize;
    fn try_intern(&self, s: &str) -> Option<InternedString>;
}

A string interner.

Required Methods

fn is_empty(&self) -> bool

Returns true if the interner contains no strings.

fn len(&self) -> usize

Returns the number of strings in the interner.

fn len_bytes(&self) -> usize

Returns the total number of bytes in the interner.

fn capacity_bytes(&self) -> usize

Returns the total number of bytes the interner can hold.

fn try_intern(&self, s: &str) -> Option<InternedString>

Attempts to intern the given string.

Returns None if the interner is full or the string cannot fit.